初步完成各类型包的测试

+++ pn-dcp/src/pn_dcp/get_req.rs
@@ -68,7 +68,7 @@ impl BlockTrait for BlockGetReq {
         len
     }
 
-    fn payload(&self) -> usize {
+    fn payload(&self) -> u16 {
         unreachable!()
     }
